Product Update feature is added

+++ frontend/src/components/ProductCard.jsx
@@ -1,17 +1,33 @@
 import { Box, Text, Heading, Image } from '@chakra-ui/react'
-import { Spacer, IconButton, HStack } from '@chakra-ui/react'
+import { Spacer, IconButton, HStack, VStack, Input } from '@chakra-ui/react'
 import { LuPencil, LuTrash2 } from "react-icons/lu";
 import React from 'react'
+import { useState } from 'react';
 import { useColorModeValue } from '@/components/ui/color-mode';
-import  {useProductStore}  from "@/store/product";
+import { useProductStore}  from "@/store/product";
+import { Button, Dialog, Portal } from "@chakra-ui/react"
+
 
 const ProductCard = ({product}) => {
     //alert("product card rendered");
-    const bgColor = useColorModeValue("white", "gray.800");
+    const bgColor = useColorModeValue("white", "gray.800"); 
 
-    const { deleteProduct} = useProductStore();
+    const {updateProduct, deleteProduct} = useProductStore();
+    const [updatedProduct, setUpdatedProduct] = useState(product);
+
+    const handleUpdatedProduct = async () => {
+        const {success, message} = await updateProduct(product._id, updatedProduct);
+        if(success){
+            console.log("msg:", message);
+        }else{
+            console.log("Product Update Failed::: ", message);
+        }
+    }
+
+    //const { deleteProduct} = useProductStore();
     const handleDeleteProduct = async () => {
-        
+
+      if (confirm("Are you sure you want to proceed?")) {        
       // Logic to handle product deletion
        const {success, message} = await deleteProduct(product._id);
          if(success){
@@ -19,6 +35,7 @@ const ProductCard = ({product}) => {
          }else{
              console.log("Product Deletion Failed::: ", message);
          }      
+      }
     }
 
   return (
@@ -33,14 +50,77 @@ const ProductCard = ({product}) => {
             <HStack spacing={2} mt={4} d="flex">     
                 <Text fontSize="lg" color="teal.600" fontWeight="bold">${product.price}</Text>
                 <Spacer />
-                <IconButton aria-label="Edit Product" variant="outline" colorScheme="blue" size="sm">
+                <Dialog.Root size="md" motionPreset="slide-in-bottom">
+  {/* The Trigger makes the button open the dialog automatically */}
+  <Dialog.Trigger asChild>
+
+    <IconButton aria-label="Edit Product" variant="outline" colorScheme="blue" size="sm" >
                     <LuPencil />
-                </IconButton>                
+                </IconButton> 
+  </Dialog.Trigger>
+  <Portal>
+    <Dialog.Backdrop />
+    <Dialog.Positioner>
+      <Dialog.Content>
+        <Dialog.Header>
+          <Dialog.Title>Edit Product</Dialog.Title>
+        </Dialog.Header>
+        <Dialog.Body>
+          
+<Box w={"full"} p={6} borderWidth={1} bg={useColorModeValue("white", "gray.800")} borderRadius="md" boxShadow="md">
+
+        <VStack spacing={4}>
+          <Input
+            type="text"
+            name="name"
+            placeholder="Product Name"
+            value={updatedProduct.name}
+            onChange={(e) => setUpdatedProduct({ ...updatedProduct, name: e.target.value })}
+          />
+          <Input
+            type="text"
+            name="description"
+            placeholder="Description"
+            value={updatedProduct.description}
+            onChange={(e) => setUpdatedProduct({ ...updatedProduct, description: e.target.value })}
+          />
+          <Input
+            type="number"
+            name="price"
+            placeholder="Price"
+            value={updatedProduct.price}
+            onChange={(e) => setUpdatedProduct({ ...updatedProduct, price: e.target.value })}
+          />
+          <Input
+            type="text"
+            name="image"
+            placeholder="Image URL"
+            value={updatedProduct.image}
+            onChange={(e) => setUpdatedProduct({ ...updatedProduct, image: e.target.value })}
+          />
+          
+        </VStack>
+      </Box>      
+
+
+        </Dialog.Body>
+        <Dialog.Footer>
+          <Dialog.ActionTrigger asChild>
+            <Button variant="outline">Cancel</Button>
+          </Dialog.ActionTrigger>
+          <Button colorScheme="blue" mt={1} onClick={handleUpdatedProduct}>Update Product</Button>
+        </Dialog.Footer>
+      </Dialog.Content>
+    </Dialog.Positioner>
+  </Portal>
+</Dialog.Root>
+                               
                 <IconButton aria-label="Delete Product" variant="outline" colorScheme="red" size="sm" onClick={handleDeleteProduct}> 
                     <LuTrash2 />
                 </IconButton>     
             </HStack>
-            </Box>
+            </Box>          
+            
     </Box>
   )
 }
